Fixed Material loading issues
The default pink/black material is now properly loaded and assigned to any meshes missing a texture (actually default assigned and overriden later). Rendering is now done with mesh materials. Added MaterialFaactory functionality to Material files to prevent cyclic dependency issues.
