+++ src/Assets/Levels/AnimatedRaptor/Materials/1/diffuse.txt
@@ -1,0 +1,1 @@
+codercat.jpg
