+++ src/Assets/Levels/AnimatedRaptor/Materials/1/emissive.txt
