+++ src/Assets/Levels/AnimatedRaptor/Meshes/1/Material.txt
@@ -1,0 +1,1 @@
+1
