+++ src/Engine/Systems/Rendering/RenderSystem.hpp
@@ -23,8 +23,6 @@
 
 class RenderSystem {
 private:
-    Material *TEMP_material;
-
     Camera *camera;
 
     // Temporary VAO to render everything for now
